style: implement SuggestionCard and page layout
This commit implements the SuggestionCard component, the EmptyState
component for when there's no feedback to display, and final adjustments
to the suggestions page layout.

Some notes on the implementation of the upvote button and comment count
components:

- All upvote buttons have an `aria-label` of "Upvote feedback" to
  indicate the primary function, an `aria-description` of "X upvotes"
  that indicates the upvote count, and an `aria-pressed` attribute that
  indicates the button's pressed status, or whether the current user has
  upvoted a feedback.

- The comment count component is implemented as a `figure` element with
  an `aria-label` of "X comments" that indicates the number of comments
  for a feedback.

Lastly, I refactored the `useSearchParams` hook and `createQueryString`
function into a custom hook that is used by the category filter and sort
select components for updating the `searchParams` as the user interacts
with the filter and sort options.