refactor: fixes and enhancements
To summarize:

1. Updating the status of any feedback revalidates the `/roadmap` page,
   which is prerendered as static content.

2. Switched to client-side data fetching for the edit feedback page to
   fix an issue where deleting a feedback abruptly shows the 404 page.

3. Posting a comment or reply updates the UI without triggering a full
   page reload via `router.refresh()`.

For #2, there was a bug where deleting a feedback causes the current
page (i.e., the edit feedback page) to rerender into a 404 page after
the `deleteFeedback` server action has returned successfully.

Fortunately, the 404 page has a link to the homepage, so the user can
continue navigating the site. But, this abrupt transition breaks the
user flow, which can be improved with a confirmation that the delete
operation has been successful.

A success modal was supposed to show after the user has confirmed the
delete operation, but this is not working because when the server action
returns, the page (which is dynamically rendered) is refreshed/reloaded,
and because the feedback doesn't exist anymore, the `notFound` route is
rendered instead.

To fix this, the edit feedback page is refactored to be statically
generated at build time, and the existing feedback data fetched on the
client via `useSWR`. This way, a successful deletion doesn't trigger a
page reload, thus giving the success modal a chance to show on the page.

Also by default, `useSWR` revalidates on focus, but since the feedback
doesn't exist once it's deleted, the UI breaks when the page refocuses
or the user switches between tabs after deleting a feedback. To fix
this, I refactored the fetcher function so that any response with status
code other than 200-299 would throw an error, which is then caught by
the `EditFeedbackFormContainer`.

I also disabled `revalidateOnFocus` for the edit feedback page.

For #3, I removed the `redirect()` after a successful invocation of the
`createComment` and `createReply` server functions. Instead I used
`router.refresh()` on the client-side to rerender the comment thread
with the newly submitted comment/reply data sans a full page reload.